Extract `printTemplateLiteral`
